\name{cgiGet2Post}
\alias{cgiGet2Post}
\alias{cgiGet2PostReset}
\title{Enable/Disable access to data generated by method GET with cgiParams}
\description{
  \code{cgiGet2Post} enables access to data generated by method GET with cgiParams. The data
  generated by method POST is not accessible after this call, unless
  \code{cgiGet2PostReset} is used.
}
\usage{
cgiGet2Post()
cgiGet2PostReset()
}
\value{
  \code{cgiGet2Post} returns (Invisible) value of non-empty \sQuote{QUERY_STRING}.  \code{NULL} if
  \sQuote{QUERY_STRING} is not defined or empty.

  \code{cgiGet2PostReset} is used for side effect and returns
  \code{NULL} only.
}
\references{
  BIOS documentation of the html module: \url{http://bioinfo.bas.roche.com:8080/bios/common/libprpi/html.hdoc}
}
\author{
  Jitao David Zhang <jitao_david.zhang@roche.com>
}
\seealso{
  \code{\link{cgiInit}}, \code{\link{cgiHeader}}, \code{\link{cgiGet2Post}}
}
\examples{
\dontrun{
cgiGet2Post()
cgiGet2PostReset()
}
}

