% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgi.R
\name{cgiGetInit}
\alias{cgiGetInit}
\title{Makes it possible to call cgiGetNext(), which will return the first field of the POSTed data, if any}
\usage{
cgiGetInit()
}
\description{
Makes it possible to call cgiGetNext(), which will return the first field of the POSTed data, if any
}

