% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{RIBIOS_CGI_TMP_DIR}
\alias{RIBIOS_CGI_TMP_DIR}
\title{CGI temporary directory}
\format{An object of class \code{character} of length 1.}
\usage{
RIBIOS_CGI_TMP_DIR
}
\description{
CGI temporary directory
}
\keyword{datasets}

