% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{RIBIOS_CGI_TMP_URL}
\alias{RIBIOS_CGI_TMP_URL}
\title{CGI temporary URL}
\format{An object of class \code{character} of length 1.}
\usage{
RIBIOS_CGI_TMP_URL
}
\description{
CGI temporary URL
}
\keyword{datasets}

