% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ronetGGI.R
\name{ronetGGI}
\alias{ronetGGI}
\title{Query gene-gene interactions (GGI) in RONET}
\usage{
ronetGGI(geneids, mode = c("in", "out", "inout", "between"),
  intTypes = NULL, sources = NULL)
}
\arguments{
\item{geneids}{A vector of Entrez GeneIDs}

\item{mode}{Interaction mode}

\item{intTypes}{Interaction types allowed}

\item{sources}{Sources allowed}
}
\description{
Query gene-gene interactions (GGI) in RONET
}
\examples{
\dontrun{
  ## query interactions between UCP1 and UCP2
  ronetGGI(c(7350, 7351), mode="between")
}

}
