% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tmpWebURL}
\alias{tmpWebURL}
\title{Get URL pointing to the temporary web file}
\usage{
tmpWebURL(tmpFile, tmpurl = RIBIOS_CGI_TMP_URL)
}
\arguments{
\item{tmpFile}{Full name of the temporary file}

\item{tmpurl}{URL root of the file}
}
\value{
Character, the URL of the temporary file
}
\description{
Get URL pointing to the temporary web file
}
\examples{
tmpFile <- tmpWebFile(pattern="file", fileext=".png")
print(tmpWebURL(tmpFile))

}
\author{
Jitao David Zhang <jitao_david.zhang@roche.com>
}
