% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tmpWebFile}
\alias{tmpWebFile}
\title{Get temporary web file}
\usage{
tmpWebFile(pattern = "file", fileext = "", tmpdir = RIBIOS_CGI_TMP_DIR)
}
\arguments{
\item{pattern}{Character, initial part of the temporary file name}

\item{fileext}{File name extesion, i.e. \dQuote{.png} (note the dot)}

\item{tmpdir}{Temporary directory}
}
\value{
Character, the full name of the temporary file
}
\description{
Get temporary web file
}
\details{
When handling web requests, it is sometimes necessary to write
 temporary files to the server file system, which is accessible by an
 URL. \code{tmpWebFile} returns a temporary file name. The URL can be
 extracted from the temp file name by \code{tmpWebURL}.

 The default temporary directory and its URL address has been adjusted to the
 BIOS installation: \code{RIBIOS_CGI_TMP_DIR} is the temporary
 directory, and \code{RIBIOS_CGI_TMP_URL} points to it. It is up to
 users to make sure that the directory and the URL points to the same
 place in case the values are explicitly specified.
}
\examples{
tmpFile <- tmpWebFile(pattern="file", fileext=".png")
print(tmpFile)
}
\author{
Jitao David Zhang <jitao_david.zhang@roche.com>
}

